+++ CHANGELOG.md
@@ -11,6 +11,7 @@
   - [#273](https://github.com/Datatamer/unify-client-python/issues/273) Attribute type spec to allow for attribute creation
   - [#219](https://github.com/Datatamer/tamr-client/issues/219) Delete a resource from collection.
   - [#277](https://github.com/Datatamer/unify-client-python/issues/277) Attribute mapping spec
+  - [#226](https://github.com/Datatamer/tamr-client/issues/226) Update published cluster configurations with put
 
   **BUG FIXES**
   - [#235](https://github.com/Datatamer/unify-client-python/issues/235) Making `AttributeCollection` retrieve attributes directly instead of by streaming